Updates to README.md
