+++ NAMESPACE
@@ -8,7 +8,8 @@ importFrom(crayon,cyan)
 importFrom(crayon,green)
 importFrom(crayon,magenta)
 importFrom(crayon,red)
-importFrom(testthat,expect_equal)
+importFrom(jsonlite,toJSON)
+importFrom(testthat,expect_equal,test_that)
 importFrom(toOrdinal,toOrdinalDate)
 importFrom(utils,data)
 importFrom(utils,packageDescription)
