+++ docs/sitemap.xml
@@ -26,7 +26,7 @@
   </url>
   <url>
     <loc>https://centerforassessment.github.io/packageSkeleton/license.html</loc>
-    <lastmod>2024-10-17T09:47:11.016Z</lastmod>
+    <lastmod>2025-05-28T08:35:52.143Z</lastmod>
   </url>
   <url>
     <loc>https://centerforassessment.github.io/packageSkeleton/presentations.html</loc>
