+++ tests/testthat.R
@@ -1,3 +1,6 @@
+# This file is part of the standard setup for testthat.
+# It is recommended that you do not modify it.
 library(testthat)
+library(packageSkeleton)
 
 test_check("packageSkeleton")
